refactor(data-seeder): give roles their own fetch module
`getUsers` fetched three documents — config/users, config/roles and
config/events — so its name described a third of what it did, and roles
were the only seed-data source without a module of their own. Nothing
about fetching or parsing them depends on users: the events fetch they
need is for the role schema, which takes event ids.

roles.ts joins locations.ts and application-config.ts as a per-source
module returning `{ seedData }`, and main() gains a fourth spread, which
is the shape it already had.

The checks that genuinely cross users and roles — a user naming an
undeclared role, and nobody holding config.update-all — stay in the
validator. Seeing every source at once is what it is for.

parse-seed-data.ts takes the three document-reading helpers the two fetch
modules now share, rather than leaving roles importing them from users.
